Fix PHPStan errors for larastan 3.x
- Add generic types to DatabaseNotificationCollection
- Type-narrow Application in CreatesApplication trait
- Cast Post::max() result to int
- Remove trivial ExampleTest

+++ tests/Feature/Models/PostTest.php
@@ -16,7 +16,7 @@ final class PostTest extends TestCase
         $user = User::factory()->createOne();
         $this->be($user);
 
-        $id = Post::max('id') + 1;
+        $id = (int) Post::max('id') + 1;
         $title = 'Some title';
         $content = 'Some content';
